inverse kinematics draft

+++ HigherLevelApps/HigherLevelApps.c/Swerve_Module.c
@@ -10,12 +10,12 @@ void Init_Swerve_Module(Swerve_Module_t *Swerve_Module, bool Azimuth_Encoder_Rev
 	Swerve_Module->Azimuth_CAN_ID = Azimuth_CAN_ID;
 }
 
-/*Optimize wheel angle so wheel doesnt have to rotate more than 90deg*/
+/*Optimize wheel angle so wheel doesn't have to rotate more than 90deg*/
 Module_State_t Optimize_Module_Angle(Module_State_t Input_State) {
 	Module_State_t Optimized_Module_State;
 	float Wheel_Angle_Delta = Input_State.Module_Angle - Measured_Wheel_Angle;
 	
-	if (fabs(Wheel_Angle_Delta) > 90.0f) {
+	if (fabsf(Wheel_Angle_Delta) > 90.0f) {
 		Optimized_Module_State.Module_Speed = -1 * Input_State.Module_Speed;
 		Optimized_Module_State.Module_Angle = Input_State.Module_Angle + 180.0f;
 	}
@@ -23,7 +23,7 @@ Module_State_t Optimize_Module_Angle(Module_State_t Input_State) {
 	return Optimized_Module_State;
 }
 
-/*Command motors to output calulated module state*/
+/*Command motors to output calculated module state*/
 void Set_Module_Output(Swerve_Module_t *Swerve_Module, Module_State_t Desired_State) {
 	Desired_State.Module_Angle = Calculate_Wrapped_Angle(Desired_State.Module_Angle);
 	Desired_State = Optimize_Module_Angle(Desired_State);
